reorganize tests to match the new structure
