+++ app/src/androidTest/java/com/delarax/dd5cv/data/characters/local/room/CharacterDAOTest.kt
@@ -1,6 +1,6 @@
-package com.delarax.dd5cv.data.database
+package com.delarax.dd5cv.data.characters.local.room
 
-import com.delarax.dd5cv.data.characters.local.room.CharacterEntity
+import com.delarax.dd5cv.data.database.AppDatabaseTest
 import kotlinx.coroutines.runBlocking
 import org.junit.Assert.assertEquals
 import org.junit.Test
