+++ app/src/androidTest/java/com/delarax/dd5cv/data/characters/local/room/ClassLevelDAOTest.kt
@@ -1,7 +1,6 @@
-package com.delarax.dd5cv.data.database
+package com.delarax.dd5cv.data.characters.local.room
 
-import com.delarax.dd5cv.data.characters.local.room.CharacterEntity
-import com.delarax.dd5cv.data.characters.local.room.ClassLevelEntity
+import com.delarax.dd5cv.data.database.AppDatabaseTest
 import kotlinx.coroutines.runBlocking
 import org.junit.Assert.assertEquals
 import org.junit.Assert.assertTrue
